requirements.txt
asgiref==3.7.2
boto3==1.28.68
botocore==1.31.68
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
Django==4.2.6
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
moto==4.2.6
packaging==23.2
psycopg-binary==3.1.12
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.23.3
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
types-PyYAML==6.0.12.12
urllib3==2.0.7
Werkzeug==3.0.0
xmltodict==0.13.0
